requirements.txt
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.9
django-phonenumber-field==1.1.0
djangorestframework==3.3.3
djangorestframework-gis==0.10.1
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.1.0
phonenumberslite==7.4.2
pickleshare==0.7.2
psycopg2==2.6.1
ptyprocess==0.5.1
pytz==2016.4
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1

